+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_SCRIPT_URL="https://script.google.com/macros/s/YOUR_SCRIPT_ID/exec"
