+++ src/App.js
@@ -1,16 +1,15 @@
+import Grading from './components/pages/Grading';
 import './styles/App.css'
-import Home from "./pages/Home";
+
 import 'bootstrap/dist/css/bootstrap.min.css';
-import Grading from './pages/Grading';
-import UpdatePointsForm from './pages/Grading';
-import ScoreTable from './pages/Grading';
+
 
 
 
 function App() {
   return (
     <>
-      <ScoreTable></ScoreTable>
+      <Grading></Grading>
       </>
       );
 }
