+++ src/components/UI/LoadingSpinner.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const LoadingSpinner = () => (
+  <div className="d-flex justify-content-center">
+    <div className="spinner-border text-primary" role="status">
+      <span className="visually-hidden">Loading...</span>
+    </div>
+  </div>
+);
+
+export default LoadingSpinner;
