+++ src/components/pages/Grading.js
@@ -1,0 +1,16 @@
+import React, { useState } from "react";
+import ScoreTable from "../fetch/ScoreTable";
+
+
+const Grading = () => {
+  const [selectedTeam, setSelectedTeam] = useState(null);
+
+  return (
+    <div >
+      <ScoreTable></ScoreTable>
+    </div>
+  );
+};
+
+export default Grading;
+
