--- src/layouts/Sidebar.js
@@ -1,10 +1,0 @@
-function Sidebar() {
-return(
-    <div>
-       <li>
-        
-       </li>
-    </div>
-)    
-}
-
