--- src/pages/Home.js
@@ -1,21 +1,0 @@
-    import Button from "../components/UI/Button"
-
-    function Home() {
-        return (
-            <div>
-            <body >
-                <div>
-                <h1>Küberküpsetus "Omlett"</h1>
-                <h2>2025</h2>
-                </div>
-                <div>
-                    <Button>HINDAJA</Button>
-                    <Button>HINDAJA</Button>
-                </div>
-            </body>
-        </div>
-        )
-       
-    }
-
-    export default Home
